[BE] refator: prisma get data 메소드 리팩토링
파일구조 front, back으로 분리
back의 model 리팩토링

user 관련 없는 테이블에 대한 get data 함수들을 폴더로 구분하여 저장
