+++ web/backend/models/albums/index.ts
@@ -1,0 +1,4 @@
+/* eslint-disable import/prefer-default-export */
+import getAlbumById from "./getById";
+
+export { getAlbumById };
