+++ web/backend/models/artists/index.ts
@@ -1,0 +1,4 @@
+/* eslint-disable import/prefer-default-export */
+import getArtistById from "./getById";
+
+export { getArtistById };
