+++ web/backend/models/dj-station/getCovers.ts
@@ -1,0 +1,8 @@
+import prisma from "../../../prisma";
+
+const getDjStationCovers = async (optObj: Object): Promise<Object> => {
+  const result = await prisma.dJStations.findMany(optObj);
+  return result;
+};
+
+export default getDjStationCovers;
