+++ web/backend/models/dj-station/index.ts
@@ -1,0 +1,4 @@
+/* eslint-disable import/prefer-default-export */
+import getDjStationCovers from "./getCovers";
+
+export { getDjStationCovers };
