+++ web/backend/models/genres/index.ts
@@ -1,0 +1,4 @@
+import getGenreById from "./getById";
+import getGenreCovers from "./getCovers";
+
+export { getGenreById, getGenreCovers };
