+++ web/backend/models/magazines/getCovers.ts
@@ -1,0 +1,8 @@
+import prisma from "../../../prisma";
+
+const getMagazineCovers = async (optObj: Object): Promise<Object> => {
+  const result = await prisma.magazines.findMany(optObj);
+  return result;
+};
+
+export default getMagazineCovers;
