+++ web/backend/models/magazines/index.ts
@@ -1,0 +1,4 @@
+import getMagazineById from "./getById";
+import getMagazineCovers from "./getCovers";
+
+export { getMagazineById, getMagazineCovers };
