+++ web/backend/models/news/getCovers.ts
@@ -1,0 +1,8 @@
+import prisma from "../../../prisma";
+
+const getNewsCovers = async (optObj: Object): Promise<Object> => {
+  const result = await prisma.news.findMany(optObj);
+  return result;
+};
+
+export default getNewsCovers;
