+++ web/backend/models/news/index.ts
@@ -1,0 +1,4 @@
+import getNewsById from "./getById";
+import getNewsCovers from "./getCovers";
+
+export { getNewsById, getNewsCovers };
