+++ web/backend/models/playlists/index.ts
@@ -1,0 +1,4 @@
+import getPlaylistById from "./getById";
+import getPlaylistCovers from "./getCovers";
+
+export { getPlaylistById, getPlaylistCovers };
