+++ web/pages/api/dj-stations/index.ts
@@ -1,5 +1,5 @@
 import { NextApiRequest, NextApiResponse } from "next";
-import prisma from "../../../prisma";
+import { getDjStationCovers } from "../../../backend/models/dj-station";
 import makeOption from "../../../backend/utils/testQuery";
 
 const handler = async (_req: NextApiRequest, res: NextApiResponse): Promise<void> => {
@@ -9,7 +9,7 @@ const handler = async (_req: NextApiRequest, res: NextApiResponse): Promise<void
     switch (method) {
       case "GET": {
         const optObj = makeOption(_req.query);
-        const result = await prisma.dJStations.findMany(optObj);
+        const result = await getDjStationCovers(optObj);
         res.status(200).json({ DJStations: result });
         break;
       }
