+++ web/pages/api/magazines/index.ts
@@ -1,5 +1,5 @@
 import { NextApiRequest, NextApiResponse } from "next";
-import prisma from "../../../prisma";
+import { getMagazineCovers } from "../../../backend/models/magazines";
 import makeOption from "../../../backend/utils/testQuery";
 
 const handler = async (_req: NextApiRequest, res: NextApiResponse): Promise<void> => {
@@ -9,7 +9,7 @@ const handler = async (_req: NextApiRequest, res: NextApiResponse): Promise<void
     switch (method) {
       case "GET": {
         const obtObj = makeOption(_req.query, "magazineType", "string");
-        const result = await prisma.magazines.findMany(obtObj);
+        const result = await getMagazineCovers(obtObj);
         res.json({ Magazines: result });
         break;
       }
